Use int for createLike scores (#44)

+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/CreatePostLike.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 internal data class CreatePostLike(
     val post_id: PostId,
-    val score: Long,
+    val score: Int,
     val auth: String,
 )
